+++ sync-for-payables/src/codatsyncpayables/models/operations/get_pull_operation.py
@@ -12,8 +12,9 @@ from typing import Optional
 @dataclasses.dataclass
 class GetPullOperationRequest:
     company_id: str = dataclasses.field(metadata={'path_param': { 'field_name': 'companyId', 'style': 'simple', 'explode': False }})
+    r"""Unique identifier for a company."""
     dataset_id: str = dataclasses.field(metadata={'path_param': { 'field_name': 'datasetId', 'style': 'simple', 'explode': False }})
-    r"""Unique ID of a dataset or pull operation."""
+    r"""Unique identifier for the dataset that completed its sync."""
     
 
 
@@ -22,11 +23,14 @@ class GetPullOperationRequest:
 @dataclasses.dataclass
 class GetPullOperationResponse:
     content_type: str = dataclasses.field()
+    r"""HTTP response content type for this operation"""
     status_code: int = dataclasses.field()
+    r"""HTTP response status code for this operation"""
     error_message: Optional[shared_errormessage.ErrorMessage] = dataclasses.field(default=None)
     r"""Your API request was not properly authorized."""
     pull_operation: Optional[shared_pulloperation.PullOperation] = dataclasses.field(default=None)
     r"""OK"""
     raw_response: Optional[requests_http.Response] = dataclasses.field(default=None)
+    r"""Raw HTTP response; suitable for custom response parsing"""